Set up API layer module and document things
